Changes requested in review
- Remove out-dated comments
- Change all span/log/metric references to "input data"
- Add comments about usage of attributes processor vs.
metric transform processor.
