+++ processor/attributesprocessor/README.md
@@ -5,17 +5,17 @@ Supported pipeline types: traces, logs, metrics.
 The attributes processor modifies attributes of a span, log, or metric. Please refer to
 [config.go](./config.go) for the config spec.
 
-This processor also supports the ability to filter and match spans/logs/metrics to determine
+This processor also supports the ability to filter and match input data to determine
 if they should be [included or excluded](#includeexclude-filtering) for specified actions.
 
 It takes a list of actions which are performed in order specified in the config.
 The supported actions are:
-- `insert`: Inserts a new attribute in spans/logs/metrics where the key does not already exist.
-- `update`: Updates an attribute in spans/logs/metrics where the key does exist.
-- `upsert`: Performs insert or update. Inserts a new attribute in spans/logs/metrics where the
-  key does not already exist and updates an attribute in spans/logs/metrics where the key
+- `insert`: Inserts a new attribute in input data where the key does not already exist.
+- `update`: Updates an attribute in input data where the key does exist.
+- `upsert`: Performs insert or update. Inserts a new attribute in input data where the
+  key does not already exist and updates an attribute in input data where the key
   does exist.
-- `delete`: Deletes an attribute from a span/log/metric.
+- `delete`: Deletes an attribute from the input data.
 - `hash`: Hashes (SHA1) an existing attribute value.
 - `extract`: Extracts values using a regular expression rule from the input key
   to target keys specified in the rule. If a target key already exists, it will
@@ -37,7 +37,7 @@ For the actions `insert`, `update` and `upsert`,
   # Key specifies the attribute to act upon.
 - key: <key>
   action: {insert, update, upsert}
-  # FromAttribute specifies the attribute from the span/log/metric to use to populate
+  # FromAttribute specifies the attribute from the input data to use to populate
   # the value. If the attribute doesn't exist, no action is performed.
   from_attribute: <other key>
 
@@ -116,11 +116,37 @@ processors:
 Refer to [config.yaml](./testdata/config.yaml) for detailed
 examples on using the processor.
 
+### Attributes Processor for Metrics vs. Metric Transform Processor
+
+Regarding metric support, these two processors have overlapping functionality. They can both do simple modifications
+of metric attribute key-value pairs. As a general rule the attributes processor has more attribute related
+functionality, while the metrics transform processor can do much more data manipulation. The attributes processor
+is preferred when the only needed functionality is overlapping, as it natively uses the official OpenTelemetry
+data model. However, if the metric transform processor is already in use or its extra functionality is necessary,
+there's no need to migrate away from it.
+
+Shared functionality
+* Add attributes
+* Update values of attributes
+
+Attribute processor specific functionality
+* delete
+* hash
+* extract
+
+Metric transform processor specific functionality
+* Rename metrics
+* Delete data points
+* Toggle data type
+* Scale value
+* Aggregate across label sets
+* Aggregate across label values
+
 ## Include/Exclude Filtering
 
 The [attribute processor](README.md) exposes
 an option to provide a set of properties of a span, log, or metric record to match against to determine
-if the span/log/metric should be included or excluded from the processor. To configure
+if the input data should be included or excluded from the processor. To configure
 this option, under `include` and/or `exclude` at least `match_type` and one of the following
 is required:
 - For spans, one of `services`, `span_names`, `attributes`, `resources`, or `libraries` must be specified
@@ -129,7 +155,7 @@ with a non-empty value for a valid configuration. The `log_names`, `expressions`
 - For logs, one of `log_names`, `attributes`, `resources`, or `libraries` must be specified with a
 non-empty value for a valid configuration. The `span_names`, `metric_names`, `expressions`, `resource_attributes`,
 and `services` fields are invalid.
-- For metrics, one of `metric_names`, `expressions`, or `resource_attributes` must be specified
+- For metrics, one of `metric_names`, `resources` must be specified
 with a valid non-empty value for a valid configuration. The `span_names`, `log_names`, and
 `services` fields are invalid.
 
@@ -161,11 +187,11 @@ attributes:
       services: [<item1>, ..., <itemN>]
 
       # resources specify an array of items to match the resources against.
-      # A match occurs if the span/log/metric resources matches at least one of the items.
+      # A match occurs if the input data resources matches at least one of the items.
       resources: [<item1>, ..., <itemN>]
 
       # libraries specify an array of items to match the implementation library against.
-      # A match occurs if the span/log/metric implementation library matches at least one of the items.
+      # A match occurs if the input data implementation library matches at least one of the items.
       libraries: [<item1>, ..., <itemN>]
 
       # The span name must match at least one of the items.
@@ -179,14 +205,6 @@ attributes:
       # The metric name must match at least one of the items.
       # This is an optional field.
       metric_names: [<item1>, ..., <itemN>]
-
-      # Expressions are used to match metrics. A match occurs if the expression matches at least
-      # one data point within a metric.
-      expressions: [<item1>, ..., <itemN>]
-
-      # Resource attributes defines a list of possible resource attributes to match metrics against.
-      # A match occurs if any resource attribute matches all expressions in this given list.
-      resource_attributes: [<item1>, ..., <itemN>]
 
       # Attributes specifies the list of attributes to match against.
       # All of these attributes must match exactly for a match to occur.
